+++ commons/src/main/java/org/apache/aurora/common/zookeeper/testing/ZooKeeperTestServer.java
@@ -19,8 +19,6 @@ import java.net.InetSocketAddress;
 
 import com.google.common.base.Preconditions;
 
-import org.apache.aurora.common.quantity.Amount;
-import org.apache.aurora.common.quantity.Time;
 import org.apache.zookeeper.server.NIOServerCnxnFactory;
 import org.apache.zookeeper.server.ServerCnxnFactory;
 import org.apache.zookeeper.server.ZooKeeperServer;
@@ -33,8 +31,6 @@ import org.apache.zookeeper.server.persistence.FileTxnSnapLog;
  * <p>This is ONLY meant to be used for testing.
  */
 public class ZooKeeperTestServer {
-
-  static final Amount<Integer, Time> DEFAULT_SESSION_TIMEOUT = Amount.of(100, Time.MILLISECONDS);
 
   private final File dataDir;
   private final File snapDir;
@@ -91,7 +87,7 @@ public class ZooKeeperTestServer {
   /**
    * Shuts down the in-process zookeeper network server.
    */
-  final void shutdownNetwork() {
+  private void shutdownNetwork() {
     if (connectionFactory != null) {
       connectionFactory.shutdown(); // Also shuts down zooKeeperServer.
       connectionFactory = null;